cc3k: Change configurability of the CC3000 module.
Optional compiling of the CC3K module is moved from mpconfigboard.h.
To enable compiling of the module, set the MICROPY_PY_CC3K variable,
either on the command line (make MICROPY_PY_CC3K=1) on in the
mpconfigport.mk file.   It is enabled by default.

Setting of the SPI and pins for the CC3K module is now done dynamically
by the arguments to wlan.init().